+++ internal/app/v1/models/schema/users.go
@@ -1,4 +1,4 @@
-package models
+package schema
 
 import (
 	"time"
